fix post video request duration type

+++ sustc-api/src/main/java/io/sustc/dto/PostVideoReq.java
@@ -30,7 +30,7 @@ public class PostVideoReq implements Serializable {
     /**
      * The video's duration (in seconds).
      */
-    private long duration;
+    private float duration;
 
     /**
      * The video's public time.
